+++ aitune/torch/tune_strategy/__init__.py
@@ -3,9 +3,17 @@
 """Torch tune strategy module."""
 
 from aitune.torch.tune_strategy.first_wins_strategy import FirstWinsStrategy
+from aitune.torch.tune_strategy.latency_budget_strategy import LatencyBudgetStrategy
 from aitune.torch.tune_strategy.max_throughput_strategy import MaxThroughputStrategy
 from aitune.torch.tune_strategy.min_latency_strategy import MinLatencyStrategy
 from aitune.torch.tune_strategy.one_backend_strategy import OneBackendStrategy
 from aitune.torch.tune_strategy.tune_strategy import TuneStrategy
 
-__all__ = ["FirstWinsStrategy", "MaxThroughputStrategy", "MinLatencyStrategy", "OneBackendStrategy", "TuneStrategy"]
+__all__ = [
+    "FirstWinsStrategy",
+    "LatencyBudgetStrategy",
+    "MaxThroughputStrategy",
+    "MinLatencyStrategy",
+    "OneBackendStrategy",
+    "TuneStrategy",
+]